requirements.txt
# langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
# streamlit
# openai==0.27.6
faiss-cpu==1.7.4
# altair==4
tiktoken==0.4.0
# uncomment to use huggingface llms
huggingface-hub==0.14.1

# uncomment to use instructor embeddings
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2

# for qdrant
qdrant-client

# for generate chart
altair>=4.2.0
commentjson==0.9.0
jsonschema
jsonref
langchain>=0.0.208
openai>=0.27.0
pandas>=1.5.0
plotly>=5.0.0
pydantic>=1.9.0
streamlit-chat>=0.0.2.2
tabulate
vegafusion[embed]>=1.0
bs4
chat2plot